Remove deps, update tests
