+++ package.json
@@ -13,7 +13,7 @@
     "probot-app"
   ],
   "scripts": {
-    "clean": "rm -rf ./lib",
+    "clean": "rm lib/**/*.js* || true && rm lib/**/*.d.ts || true",
     "build": "tsc -p tsconfig.json",
     "dev": "run-p dev:build dev:server",
     "dev:build": "yarn build -w",
